refactor(fieldmap): move plot state into plot grid context
